fix stock

+++ DATN-Theme-main/src/admin/adminFlashSales/flashsale.jsx
@@ -4,6 +4,9 @@ import { jwtDecode } from 'jwt-decode';
 import Cookies from 'js-cookie';
 import { ToastContainer, toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
+import { confirmAlert } from 'react-confirm-alert'; // Nhập confirmAlert
+import 'react-confirm-alert/src/react-confirm-alert.css'; // Nhập CSS của confirmAlert
+
 
 const FlashSaleManagementPage = () => {
   const [isFormVisible, setFormVisible] = useState(false);
@@ -236,24 +239,41 @@ const FlashSaleManagementPage = () => {
     }
   };
 
-  const handleDeleteProductFlashsale = async (productId, flashsaleId) => {
-    try {
-      const response = await axios.delete(`http://localhost:8080/api/admin/product-flashsale/delete/${productId}/${flashsaleId}`, {
-        headers: {
-          Authorization: `Bearer ${token}`, // Đảm bảo token đã được định nghĩa đúng
-        },
-      });
+const handleDeleteProductFlashsale = (productId, flashsaleId) => {
+  confirmAlert({
+    title: 'Xác nhận xóa',
+    message: 'Bạn có chắc chắn muốn xóa sản phẩm này khỏi Flash Sale?',
+    buttons: [
+      {
+        label: 'Có',
+        onClick: async () => {
+          try {
+            const response = await axios.delete(`http://localhost:8080/api/admin/product-flashsale/delete/${productId}/${flashsaleId}`, {
+              headers: {
+                Authorization: `Bearer ${token}`, // Đảm bảo token đã được định nghĩa đúng
+              },
+            });
 
-      if (response.status === 200) {
-        toast.success(response.data); // Hiển thị thông báo thành công
-        // Cập nhật danh sách sản phẩm trong flashsale sau khi xóa
-        await fetchProductFlashSales(); // Gọi lại API để lấy danh sách cập nhật
-      }
-    } catch (error) {
-      console.error('Lỗi khi xóa ProductFlashsale:', error);
-      toast.error('Có lỗi xảy ra khi xóa ProductFlashsale.');
-    }
-  };
+            if (response.status === 200) {
+              toast.success(response.data); // Hiển thị thông báo thành công
+              // Cập nhật danh sách sản phẩm trong flashsale sau khi xóa
+              await fetchProductFlashSales(); // Gọi lại API để lấy danh sách cập nhật
+            }
+          } catch (error) {
+            console.error('Lỗi khi xóa ProductFlashsale:', error);
+            toast.error('Có lỗi xảy ra khi xóa ProductFlashsale.');
+          }
+        },
+      },
+      {
+        label: 'Không',
+        onClick: () => {
+          toast.info('Hành động xóa đã bị hủy.');
+        },
+      },
+    ],
+  });
+};
